// src/constants/permission.enum.ts
export enum PermissionName {
    CREATE = 'create',
    READ = 'read',
    UPDATE = 'update',
    DELETE = 'delete'
}

export enum PermissionResource {
    USERS = 'users',
    PRODUCTS = 'products',
    CATEGORIES = 'categories',
    INVENTORIES = 'inventories',
    ROLES = 'roles',
    PERMISSIONS = 'permissions',
    ORDERS = 'orders'
}